docs: temporary config which best appoximates the version2 published by the authors
I took the config file published by the authors with pretrained model (version2) and tried to change all parameters inside the amass.yaml to the same values as in version2 config file. I had to turn off the flip feature, because I am not able to run the code with it turned on. Definitely we need to replicate the results with more diligence